Switch to a type-level representation of metadata that is not a GADT.
This makes some things simpler, hopefully, because it means that the
kind of the type-level metadata is always the same.

Whether convenience is better or not remains to be seen, but at least
some things should become easier. Safety is retained, because the
whole metadata value is of course statically known.